Added ProtoPics as a directory to hold pictorial prototypes of the UI LayoutManager.png is in an early stage of showing the UI. It has the garden bed in brown, and a list of plants on the right hand side. I have placed some plants. The brown background and white circles on the far right were just for image editing.
As previously discussed, the garden bed should occupy as much screen space as possible most of the time.
The list (or dialogue) should only appear when the user wants to select a plant.
The user may have to press a dedicated button to bring up the list, or the user might have to press on the garden bed to bring up a dialogue, or <your favorite UI experience> ...
I have yet to reflect these ideas in the protoype.